finished modules of the server side js files
