+++ Modules/renderLogin.js
@@ -1,0 +1,5 @@
+function renderLogin(req, res) {
+    res.render('complete/login', { 'accountCreated': false, 'failed': false, 'loggedIn': false });
+  }
+
+  module.exports = renderLogin;
