+++ views/partials/navbar.ejs
@@ -7,7 +7,7 @@
         <p class="eco-credit-display">Eco-Credit: <%=ecoscore%> </p>
       <% } %>
 
-      <a href="/account/view/?username=<%= user %>&id=<%= id%>" class="nav-link">
+      <a href="/account/view/?username=<%= user %>&id=<%= id %>" class="nav-link">
         <p>View Account</p>
       </a>
